requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.9.1
    # via twilio
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==23.2.0
    # via aiohttp
audiosegment==0.23.0
    # via -r requirements.in
backports-functools-lru-cache==2.0.0
    # via indic-transliteration
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.8
    # via
    #   flask
    #   gtts
    #   typer
coverage[toml]==7.9.1
    # via pytest-cov
deprecated==1.2.18
    # via limits
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-limiter
flask-limiter==3.5.0
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.24.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.166.0
    # via google-generativeai
google-auth==2.38.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.4
    # via -r requirements.in
googleapis-common-protos==1.69.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
gtts==2.5.4
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
huggingface-hub==0.30.2
    # via
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   requests
    #   yarl
indic-transliteration==2.3.69
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.3
    # via
    #   flask
    #   torch
limits==5.4.0
    # via flask-limiter
llvmlite==0.44.0
    # via numba
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.6.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
multidict==6.6.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numba==0.61.0
    # via openai-whisper
numpy==1.26.4
    # via
    #   audiosegment
    #   numba
    #   openai-whisper
    #   pandas
    #   transformers
openai-whisper==20240930
    # via -r requirements.in
ordered-set==4.1.0
    # via flask-limiter
packaging==23.2
    # via
    #   huggingface-hub
    #   limits
    #   pytest
    #   transformers
pandas==2.2.3
    # via -r requirements.in
playsound==1.3.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.4
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.9.2
    # via google-generativeai
pydantic-core==2.23.4
    # via pydantic
pydub==0.25.1
    # via audiosegment
pygments==2.17.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via twilio
pyparsing==3.1.2
    # via httplib2
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==6.2.1
    # via -r requirements.in
pytest-mock==3.14.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2024.9.11
    # via
    #   indic-transliteration
    #   tiktoken
    #   transformers
requests==2.31.0
    # via
    #   google-api-core
    #   gtts
    #   huggingface-hub
    #   tiktoken
    #   transformers
    #   twilio
rich==13.9.3
    # via
    #   flask-limiter
    #   typer
roman==5.0
    # via indic-transliteration
rsa==4.9
    # via google-auth
safetensors==0.5.3
    # via transformers
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
speechrecognition==3.14.2
    # via -r requirements.in
sympy==1.13.1
    # via torch
tiktoken==0.8.0
    # via openai-whisper
tokenizers==0.21.1
    # via transformers
toml==0.10.2
    # via indic-transliteration
torch==2.6.0
    # via openai-whisper
tqdm==4.66.5
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai-whisper
    #   transformers
transformers==4.51.1
    # via -r requirements.in
twilio==9.6.3
    # via -r requirements.in
typer==0.16.0
    # via indic-transliteration
typing-extensions==4.12.2
    # via
    #   flask-limiter
    #   google-generativeai
    #   huggingface-hub
    #   limits
    #   pydantic
    #   pydantic-core
    #   speechrecognition
    #   torch
    #   typer
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.0
    # via requests
webrtcvad==2.0.10
    # via audiosegment
werkzeug==3.1.3
    # via flask
word2number==1.1
    # via -r requirements.in
wrapt==1.17.2
    # via deprecated
yarl==1.20.1
    # via aiohttp
